fixup! bench: add seriesMetadata unused-metrics benchmarks and raise page size cap
